Refactor to add ResourceLink handler
Decided it's better for the handlers to know everything about whether
they are valid, and to deal with the iterating-through-image-urls
process. Fetcher's job is now only to 1) know about available handlers,
and 2) iterate through them until it gets some text (or runs out).

The handlers are still pretty gross inside, BUT they now share a common
(and small) public interface, with nearly everything moved to private
methods.

This makes it possible to easily add new handlers for newly discovered
edge cases. The ResourceLink is the best where available, because it
uses LOC's own knowledge of where to find fulltext files, rather than
reverse-engineering it.