Fetch fulltext by scraping URL
This swaps the Fetcher.full_text_from_url implementation for one that
just finds the download form on the web site and pulls the fulltext
URL. On net this will be fewer queries than the API.

This almost always works to fetch fulltext; the errors almost all look
intermittent, not structural.

However, the XML parsing is incorrect.